Update media modal periodically for new uploads
When the media attachment meta is showing the analysing spinner we can set WordPress up to periodically refresh the attachment model. This will update the UI once the label data has been returned or if an error occurs.

Any modifications to the fields like alt text etc do not get wiped out by doing this.